# Refactor linalg tests to make them faster. Add timings to tests.