--- a/test/testdefs.jl
+++ b/test/testdefs.jl
@@ -1,8 +1,9 @@
 using Base.Test
 
 function runtests(name)
-    println("     \033[1m*\033[0m \033[31m$(name)\033[0m")
-    Core.include("$name.jl")
+    @printf("     \033[1m*\033[0m \033[31m%-20s\033[0m", name)
+    tt = @elapsed Core.include("$name.jl")
+    @printf(" in %6.2f seconds\n", tt)
     nothing
 end