@andreasnoack 2015-01-19
I'll merge this to get the test speedup. I we don't like the timings in the output it can be removed later.