@staticfloat 2015-01-19
Oooh, I actually do really like the timings in the output.  Nice work, Andreas!