@tkelman 2015-01-19
+ a lot. Though I wonder why the color codes aren't working when tests run in serial?